+++ packages/lib/helpers.js
@@ -1,3 +1,5 @@
+import compose from 'crocks/helpers/compose';
+import filter from 'crocks/pointfree/filter';
 const scriptSelector = 'script[type="application/graphql"]';
 
 export const isGraphQLScript = el =>
@@ -10,3 +12,10 @@ export const getGraphQLScriptChild = el =>
 
 export const replace = (...args) => str =>
   typeof str === 'string' ? str.replace(...args) : str;
+
+const valueIsUndefined = ([, v]) => v !== undefined;
+export const stripUndefinedValues = compose(
+  Object.fromEntries,
+  filter(valueIsUndefined),
+  Object.entries,
+);