+++ packages/lib/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/apollo-elements/apollo-elements/compare/@apollo-elements/lib@0.0.6...@apollo-elements/lib@1.0.0) (2019-04-03)
+
+
+### Bug Fixes
+
+* **mixins:** allow default options in queries ([4a8895e](https://github.com/apollo-elements/apollo-elements/commit/4a8895e))
+
+
+
+
+
 ## [0.0.6](https://github.com/apollo-elements/apollo-elements/compare/@apollo-elements/lib@0.0.5...@apollo-elements/lib@0.0.6) (2019-03-01)
 
 **Note:** Version bump only for package @apollo-elements/lib